+++ src/user-event/setup/setup.ts
@@ -10,8 +10,7 @@ import { paste } from '../paste';
 import type { PressOptions } from '../press';
 import { longPress, press } from '../press';
 import type { ScrollToOptions } from '../scroll';
-import { scrollTo } from '../scroll';
-import { pullToRefresh } from '../scroll/pull-to-refresh';
+import { pullToRefresh, scrollTo } from '../scroll';
 import type { TypeOptions } from '../type';
 import { type } from '../type';
 import { wait } from '../utils';
@@ -150,7 +149,7 @@ export interface UserEventInstance {
   paste: (instance: TestInstance, text: string) => Promise<void>;
 
   /**
-   * Simlate user scorlling a given `ScrollView`-like element.
+   * Simulate user scrolling a given `ScrollView`-like element.
    *
    * Supported components: ScrollView, FlatList, SectionList
    *